# forum/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planinarski Forum API",
        "description": "API dokumentacija za Planinarski Forum aplikaciju",
        "contact": {
            "email": "admin@planinarskiforum.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
...
        "/api/themes": {
            "get": {
                "tags": [
                    "Themes"
                ],
                "summary": "Lista svih tema",
                "description": "Vraća listu svih tema sa brojem postova",
                "operationId": "3af0ee5eb0876976d8f366507e93e82a",
                "responses": {
                    "200": {
                        "description": "Uspešno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "themes": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nema tema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "No themes found."
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Themes"
                ],
                "summary": "Kreiranje nove teme",
                "description": "Kreira novu temu (samo admin)",
                "operationId": "88a51c58f980c0ba8846ebadd3a9f87a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Naziv teme",
                                        "type": "string",
                                        "example": "Zimsko planinarenje"
                                    },
                                    "description": {
                                        "description": "Opis teme",
                                        "type": "string",
                                        "example": "Sve o zimskom planinarenju"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tema kreirana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Theme created successfully"
                                        },
                                        "theme": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Nije admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Only admins can create themes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Greška pri validaciji"
                    },
                    "401": {
                        "description": "Nije autentifikovan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/themes/{theme}": {
            "get": {
                "tags": [
                    "Themes"
                ],
                "summary": "Pojedinačna tema",
                "description": "Vraća detalje jedne teme sa brojem postova",
                "operationId": "b26fa44f6f494cf0cbcc65d38b232cbe",
                "parameters": [
                    {
                        "name": "theme",
                        "in": "path",
                        "description": "ID teme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "theme": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tema nije pronađena"
                    }
                }
            },
            "put": {
                "tags": [
                    "Themes"
                ],
                "summary": "Ažuriranje teme",
                "description": "Ažurira postojeću temu (samo admin)",
                "operationId": "2279de886d59625c857567cc236b8a50",
                "parameters": [
                    {
                        "name": "theme",
                        "in": "path",
                        "description": "ID teme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Naziv teme",
                                        "type": "string",
                                        "example": "Zimsko planinarenje"
                                    },
                                    "description": {
                                        "description": "Opis teme",
                                        "type": "string",
                                        "example": "Sve o zimskom planinarenju"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tema ažurirana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Theme updated successfully"
                                        },
                                        "theme": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Nije admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Only admins can update themes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Greška pri validaciji"
                    },
                    "401": {
                        "description": "Nije autentifikovan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Themes"
                ],
                "summary": "Brisanje teme",
                "description": "Briše temu (samo admin)",
                "operationId": "0ad1a88ab068d11a34b3572bdbb2cbfa",
                "parameters": [
                    {
                        "name": "theme",
                        "in": "path",
                        "description": "ID teme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tema obrisana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Theme deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Nije admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Only admins can delete themes"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nije autentifikovan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/themes/{theme}/posts": {
            "get": {
                "tags": [
                    "Themes"
                ],
                "summary": "Postovi u temi",
                "description": "Vraća listu postova za određenu temu",
                "operationId": "dfca4c3274dfe8200e33cf71636e38ff",
                "parameters": [
                    {
                        "name": "theme",
                        "in": "path",
                        "description": "ID teme",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "theme": {
                                            "type": "object"
                                        },
                                        "posts": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tema nije pronađena"
                    }
                }
            }
        },
...
    },
...
}